#3. #2. Made some file fixes and compiled.
I moved the files from the folder within a folder and made some small edits.
