--- RozemicBanking/RozemicBanking/CustomerInfo.cpp
@@ -1,11 +1,0 @@
-
-	#include <iostream>
-	#include <string>
-	#include "CustomerInfo.h"
-		
-	int main()
-	{
-
-		CustomerInfo addcustomer("John", "Doe","3001 Mercer University Dr", "Atlanta", "GA", "30341");
-		std::cout << addcustomer << '\n';
-	}
